YouTube-Subscription-Transfer/requirements.txt
google-auth==2.22.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
google-api-python-client==2.92.0